+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/OspfInternalRoute.java
@@ -10,7 +10,7 @@ public abstract class OspfInternalRoute extends OspfRoute {
   protected OspfInternalRoute(
       Prefix network,
       NextHop nextHop,
-      int admin,
+      long admin,
       long metric,
       long area,
       long tag,
